+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 Version 1.5 (not yet released)
-
+------------------------------
 
 
 Version 1.4